Create Microsoft codility interview: Find the index of a character that occurs in a pair of string in the array of strings
Given an array S consisiting of N integers. Every string is the same length. your task is to find a pair of string that contain particular character in the same index position.

The solution should return an array that contain the index of each the pair of strings, and the indes of the character found in the same position in each string.
e.g S=['abc', 'abd', 'bbd']

the function should return [0,2,1] or [1,2,2]
Because 'abc' and 'bbd' are valid pairs where 'b' occurs in the same position of 1 for both strings
And also 'abd' and 'bbd' are valid pairs because 'd' occurs in the same position of 2 for both strings.
The function can return either of the two arrays.